refactor: Extract filter logic to a new function
- Added `filterDataBySearchParams` function for filtering data based on search parameters.
- Moved filtering logic from route handler to the new function for better code organization.
- Updated imports to use the new function.
